requirements.txt
# Core dependencies for Six Worker
# Cloudflare Workers Python runtime dependencies

# Database connectivity
psycopg2-binary>=2.9.5

# Data processing and utilities
uuid>=1.30
python-dateutil>=2.8.2

# LLM integration for test data generation
anthropic>=0.25.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.1

# Data validation and processing
pydantic>=2.0.0
jsonschema>=4.17.0

# Optional: For advanced data generation
faker>=19.0.0